Criado tradutor de instruções

+++ Montador/Assembler.cpp
@@ -371,7 +371,7 @@ void Assembler::ValidateAndWriteInstruction(const InstructionInfo * info, const 
 		positionCount++;
 	}
 
-	generator.WriteInstruction((int)info->opCode, arguments);
+	generator.WriteInstruction(info->opCode, arguments);
 }
 
 bool Assembler::TryCalculateOperandRealValue(string operand, int & extractedValue, SymbolInfo** symbol, string& symbolName) {
